Factory pattern
From :
http://www.tutorialspoint.com/design_pattern/design_pattern_tutorial.pdf


Factory pattern is one of most used design pattern in Java. This type of
design pattern 
comes under creational pattern as this pattern provides one of the best
ways to create an object. 
In Factory pattern, we create object without exposing the creation logic
to the client and refer to 
newly created object using a common interface.